pyproject.toml
[project]
name = "resume-parser"
version = "1"
description = "Resume parser microservice"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "3.11"
dependencies = [
    "pdfplumber (>=0.11.5,<0.12.0)",
    "spacy (>=3.8.4,<4.0.0)",
    "python-docx (>=1.1.2,<2.0.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pika (>=1.3.2,<2.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "pyspellchecker (>=0.8.2,<0.9.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.packages]]
include = "src"